refactor : filter/test 패키지 삭제, 파일 위치 이동
